fix(shared-influx-pieces): move utils files and rename AnalyticsListener to AnalyticsListener
