+++ packages/shared-influx-pieces/src/lib/enum.ts
