+++ packages/shared-influx-pieces/src/register.ts
@@ -1,7 +1,7 @@
 import './listeners/_load';
 
 import { initializeInflux } from './index.js';
-import { registerSchedule } from './lib/utils/schedule';
+import { registerSchedule } from './lib/schedule';
 
 void registerSchedule();
 void initializeInflux();
